Moved proto types
