+++ src/utils/transport/types/messageUpdate/workspaceUpdated.ts
@@ -1,4 +1,4 @@
-import { Workspace } from '../../../types';
+import { Workspace } from '../../../../types';
 import { NewMessage } from 'ctproto';
 
 /**
